fix(emulator): implement proper deferred (pending) auto wrap
We should not immediately advance to the next line on automargin,
but only if we receive another character.  This pending state can
be reset by a number of values.

(As an interesting note: the famous bug with scrolling when writing
to the bottom right corner was caused by implementations that did not
properly pend the automatic wrapping. Correct implementations can
write to the final corner without this behavior, even with automatic
margin wrapping enabled. A number of buggy real-world implementations
exist.)